% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{cmap.knitr}
\alias{cmap.knitr}
\title{Knitr color map}
\usage{
cmap.knitr(n)
}
\arguments{
\item{n}{integer vector of color indices.}
}
\value{
RGB colors.
}
\description{
8 distinct colors compatible with knitr's default colors.
}
