% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap.R
\name{cmap.seaborn}
\alias{cmap.seaborn}
\title{Seaborn color map}
\usage{
cmap.seaborn(n)
}
\arguments{
\item{n}{integer vector of color indices.}
}
\value{
RGB colors.
}
\description{
10 distinct colors from https://seaborn.pydata.org/.
}
