% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{hist}
\alias{hist}
\title{Replacement to hist()}
\usage{
hist(
  x,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  alpha = 0.6,
  cmap = NULL,
  col = 1,
  border = rgb(0, 0, 0, 0),
  freq = F,
  breaks = NULL,
  yaxt = "n",
  pch = 20,
  bty = "n",
  lwd = 0.75,
  tcl = -0.15,
  mgp = c(2, 0.25, 0),
  cex.axis = 0.8,
  family = "serif",
  axes = T,
  showmean = F,
  xmark = NULL,
  xmark.label = NULL,
  ...
)
}
\description{
Replacement to hist()
}
