% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Pretty replacement to plot()}
\usage{
plot(
  x,
  y = NULL,
  xlab = NULL,
  ylab = NULL,
  alpha = 0.6,
  cmap = cmap.seaborn,
  col = 1,
  pch = 20,
  cex.axis = 0.8,
  cex = 1,
  family = "serif",
  xmark = NULL,
  xmark.label = NULL,
  ymark = NULL,
  ymark.label = NULL,
  axes = TRUE,
  ...
)
}
\arguments{
\item{x}{coordinates for the plot.}

\item{y}{y-axis coordinates for the plot (optional).}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{alpha}{opacity of the plot's points or lines.}

\item{cmap}{color map to use.}

\item{col}{colors.}

\item{pch}{Point style}

\item{cex.axis}{Scale of the plot axes.}

\item{cex}{Scale of the plot features.}

\item{family}{Font family.}

\item{xmark}{Optional mark on the x-axis.}

\item{xmark.label}{Optional mark label on the x-axis.}

\item{ymark}{Optional mark on the y-axis.}

\item{ymark.label}{Optional mark label on the y-axis.}

\item{axes}{whether or not to show axes.}
}
\description{
Pretty replacement to plot()
}
