% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Pretty plot}
\usage{
plot(x, y, xlab = NULL, ylab = NULL, alpha = 0.6, cmap = NULL,
  ptscale = NULL, col = 1, pch = 20, bty = "n", tcl = -0.15,
  mgp = c(2, 0.25, 0), cex.axis = 0.8, cex = 1, ...)
}
\arguments{
\item{x}{The coordinates of points of the plot. Alternatively, a single
plotting structure, function or any R object with a plot method can
be provided.}

\item{y}{The y coordinates of points in the plot, optional if x is an
appropriate structure.}

\item{alpha}{Adjusts the opacity of the color specified by the parameter \code{col}.}

\item{cmap}{Map to be applied to the parameter \code{col}.}

\item{ptscale}{Symbol size adjustment in log-scale (relatively to cex). Positive
values represent an increased size, while negative value represent
a decreased size.}

\item{col}{A specification for the default plotting color. By default, a value
of 0 represents black. Other positive integer values are mapped to
colors by the default color map.}
}
\description{
Wrapper to R's \code{\link{graphics::plot}} function. Overrides
    default behavior for better plots.
}
